requirements.txt
# Runtime
ollama>=0.4.0
pyyaml>=6.0
python-docx>=1.1
requests>=2.32
beautifulsoup4>=4.12

# Dev / test
pytest>=8.0
pytest-mock>=3.14
ruff>=0.9.0
